+++ src/od-fs/include/uae/uae.h
@@ -145,7 +145,7 @@ typedef struct _RenderData {
     char line[AMIGA_MAX_LINES];
     int flags;
     void *(*grow)(int width, int height);
-    int refresh_rate;
+    double refresh_rate;
     int bpp;
 } RenderData;
